HPCC-18869 Refactoring of RowCursor prior to integration into Roxie
